+++ queries/hashtags.json
@@ -1,4 +1,4 @@
-[{"limit": 100,
+[{"limit": 1000,
   "kinds": [
     1
   ],
